+++ application/modules/content/controllers/IndexController.php
@@ -65,7 +65,7 @@ class Content_IndexController extends Zend_Controller_Action
      */
     public function indexAction()
     {
-        $model_sites = new Dlayer_Model_Site();
+        $model_sites = new Dlayer_Model_Admin_Site();
         $model_pages = new Dlayer_Model_Admin_Page();
         
         $pages = $model_pages->pages($this->site_id);
@@ -95,7 +95,7 @@ class Content_IndexController extends Zend_Controller_Action
      */
     private function controlBar($identity_id, $site_id)
     {
-        $model_sites = new Dlayer_Model_Site();
+        $model = new Dlayer_Model_Admin_Site();
 
         $control_bar_buttons = array(
             array(
@@ -114,7 +114,7 @@ class Content_IndexController extends Zend_Controller_Action
             array(
                 'name' => 'Your websites',
                 'class' => 'default',
-                'buttons' => $model_sites->sitesForControlBar($identity_id, $site_id)
+                'buttons' => $model->sitesForControlBar($identity_id, $site_id)
             )
         );