+++ application/modules/dlayer/controllers/AdminController.php
@@ -1,13 +1,13 @@
 <?php
 
 /**
- * Admin controller for Content manager
+ * Admin controller for Dlayer
  *
  * @author Dean Blackborough <dean@g3d-development.com>
  * @copyright G3D Development Limited
  * @license https://github.com/Dlayer/dlayer/blob/master/LICENSE
  */
-class Content_AdminController extends Zend_Controller_Action
+class Dlayer_AdminController extends Zend_Controller_Action
 {
     /**
      * Type hinting for action helpers
@@ -23,11 +23,6 @@ class Content_AdminController extends Zend_Controller_Action
     private $form;
 
     /**
-     * @var Dlayer_Session_Content
-     */
-    private $session;
-
-    /**
      * @var array Nav bar items
      */
     private $nav_bar_items = array(
@@ -35,11 +30,6 @@ class Content_AdminController extends Zend_Controller_Action
             'uri' => '/dlayer/index/home',
             'name' => 'Dlayer Demo',
             'title' => 'Dlayer.com: Web development simplified'
-        ),
-        array(
-            'uri' => '/content/index/index',
-            'name' => 'Content Manager',
-            'title' => 'Content Manager'
         ),
         array(
             'uri' => 'http://www.dlayer.com/docs/',
@@ -63,54 +53,6 @@ class Content_AdminController extends Zend_Controller_Action
 
         $this->site_id = $session_dlayer->siteId();
         $this->identity_id = $session_dlayer->identityId();
-
-        $this->session = new Dlayer_Session_Content();
-    }
-
-    /**
-     * Add a new content page
-     *
-     * @return void
-     */
-    public function newAction()
-    {
-        $model_sites = new Dlayer_Model_Admin_Site();
-
-        $this->form = new Dlayer_Form_Admin_Page('/content/admin/new', $this->site_id);
-
-        if ($this->getRequest()->isPost()) {
-            $this->handleAddPage();
-        }
-
-        $this->view->form = $this->form;
-        $this->view->site = $model_sites->site($this->site_id);
-
-        $this->controlBar($this->identity_id, $this->site_id);
-
-        $this->_helper->setLayoutProperties($this->nav_bar_items, '/content/index/index', array('css/dlayer.css'),
-            array(), 'Dlayer.com - Content Manager: New page');
-    }
-
-    /**
-     * Handle add content page, if successful the user is redirected after the ids for the new page has been set in
-     * the session
-     *
-     * @return void
-     */
-    private function handleAddPage()
-    {
-        $post = $this->getRequest()->getPost();
-
-        if ($this->form->isValid($post)) {
-            $model = new Dlayer_Model_Admin_Page();
-            $page_id = $model->savePage($this->site_id, $post['name'], $post['title'], $post['description']);
-
-            if ($page_id !== false) {
-                $this->session->clearAll(true);
-                $this->session->setPageId($page_id);
-                $this->redirect('/content');
-            }
-        }
     }
 
     /**
@@ -132,9 +74,9 @@ class Content_AdminController extends Zend_Controller_Action
                 'name' => 'Dashboard'
             ),
             array(
-                'uri'=>'/content/admin/new',
+                'uri'=>'/dlayer/admin/new',
                 'class' => 'primary',
-                'name'=>'New page'
+                'name'=>'New website'
             )
         );
 
@@ -150,47 +92,91 @@ class Content_AdminController extends Zend_Controller_Action
     }
 
     /**
-     * Edit the selected content page
+     * Create a new website, at the moment a user only needs to define the name they want to use, later there
+     * will be an updated to set additional data as well as some Dlayer options
+     *
+     * @return void
+     */
+    public function newAction()
+    {
+        $this->_helper->authenticate();
+
+        $this->form = new Dlayer_Form_Admin_Site('/dlayer/admin/new');
+
+        if ($this->getRequest()->isPost()) {
+            $this->handleAdd();
+        }
+
+        $model_sites = new Dlayer_Model_Admin_Site();
+        $this->view->site = $model_sites->site($this->site_id);
+        $this->view->form = $this->form;
+
+        $this->controlBar($this->identity_id, $this->site_id);
+
+        $this->_helper->setLayoutProperties($this->nav_bar_items, '/dlayer/index/home', array('css/dlayer.css'),
+            array(), 'Dlayer.com - New website');
+    }
+
+    /**
+     * Handle add web site, if successful the user is redirected back to app root, site is not automatically activate
+     * done by a different action because it is more involved than just setting an id
+     *
+     * @return void
+     */
+    private function handleAdd()
+    {
+        $post = $this->getRequest()->getPost();
+
+        if ($this->form->isValid($post)) {
+            $model = new Dlayer_Model_Admin_Site();
+            $site_id = $model->saveSite($post['name']);
+
+            if ($site_id !== false) {
+                $this->redirect('/dlayer/index/home');
+            }
+        }
+    }
+
+    /**
+     * Edit the selected web site
      *
      * @return void
      */
     public function editAction()
     {
-        $model_sites = new Dlayer_Model_Admin_Site();
+        $this->_helper->authenticate();
 
-        $this->form = new Dlayer_Form_Admin_Page('/content/admin/edit', $this->site_id,
-            $this->session->pageId());
+        $this->form = new Dlayer_Form_Admin_Site('/dlayer/admin/edit', $this->site_id);
 
         if ($this->getRequest()->isPost()) {
-            $this->handleEditPage();
+            $this->handleEdit();
         }
 
+        $model = new Dlayer_Model_Admin_Site();
+        $this->view->site = $model->site($this->site_id);
         $this->view->form = $this->form;
-        $this->view->site = $model_sites->site($this->site_id);
 
-        $session_dlayer = new Dlayer_Session();
-        $this->controlBar($session_dlayer->identityId(), $this->site_id);
+        $this->controlBar($this->identity_id, $this->site_id);
 
-        $this->_helper->setLayoutProperties($this->nav_bar_items, '/content/index/index', array('css/dlayer.css'),
-            array(), 'Dlayer.com - Edit content page');
+        $this->_helper->setLayoutProperties($this->nav_bar_items, '/dlayer/index/home', array('css/dlayer.css'),
+            array(), 'Dlayer.com - Edit website');
     }
 
     /**
-     * Handle edit content page, if successful the user is redirected back to Content manager root
+     * Handle edit web site, if successful the user is redirected back to app root
      *
      * @return void
      */
-    private function handleEditPage()
+    private function handleEdit()
     {
         $post = $this->getRequest()->getPost();
 
         if ($this->form->isValid($post)) {
-            $model = new Dlayer_Model_Admin_Page();
-            $page_id = $model->savePage($this->site_id, $post['name'], $post['title'], $post['description'],
-                $this->session->pageId());
+            $model = new Dlayer_Model_Admin_Site();
+            $site_id = $model->saveSite($post['name'], $this->site_id);
 
-            if ($page_id !== false) {
-                $this->redirect('/content');
+            if ($site_id !== false) {
+                $this->redirect('/dlayer/index/home');
             }
         }
     }